third_party/WebKit/LayoutTests/webmessaging/MessageEvent_onmessage_infinite_loop.html

<!DOCTYPE html>
<html>
<head>
<title>MessageEvent: onmessage infinite loop</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body>
<script>
'use strict';

// looper will call port1's postmessage from port2's onmessage to create many
// message events. These events should not prevent window.setTimeout's callback
// from ever being called.
async_test(function(t) {
  var onmessage_seen = 0;
  var looper = function() {
    var channel = new MessageChannel();
    var callback;
    channel['port1'].onmessage = function() {
      onmessage_seen++;
      if (onmessage_seen < 100) {
        // calls port2's postMessage again.
        callback();
      }
    };    
    return function(cb) {
      callback = cb;
      channel['port2'].postMessage(0);
    };
  }();
  function run() {
    looper(run);
  }
  run();

  window.setTimeout(function() {
    t.step(function() {
      assert_less_than(onmessage_seen, 100,
                      "setTimeout was not called until after the onmessage " +
                      "events stopped.");
    });
    t.done();
  }, 0);
}, "onmessage calling source postMessage")
</script>
